+++ packages/core/src/instrumentations/types.ts
@@ -3,4 +3,5 @@ export interface Instrumentation {
   readonly version: string;
 
   initialize(): void;
+  shutdown(): void
 }