+++ packages/web/src/instrumentations/console/index.ts
@@ -9,15 +9,19 @@ export class ConsoleInstrumentation extends BaseInstrumentation {
 
   readonly version = VERSION;
   readonly name = '@grafana/agent-web:instrumentation-console';
+  private originalMethods: Map<LogLevel, typeof console.info> = new Map()
 
   constructor(private options: ConsoleInstrumentationOptions = {}) {
     super();
   }
 
   initialize() {
+    this.unpatch()
     allLogLevels
       .filter((level) => !(this.options.disabledLevels ?? ConsoleInstrumentation.defaultDisabledLevels).includes(level))
       .forEach((level) => {
+        /* eslint-disable-next-line no-console */
+        this.originalMethods.set(level, console[level]);
         /* eslint-disable-next-line no-console */
         console[level] = (...args) => {
           try {
@@ -28,5 +32,17 @@ export class ConsoleInstrumentation extends BaseInstrumentation {
           }
         };
       });
+  }
+
+  private unpatch(): void {
+   this.originalMethods.forEach((fn, level) => {
+    /* eslint-disable-next-line no-console */
+    console[level] = fn;
+   })
+   this.originalMethods.clear()
+  }
+
+  shutdown(): void {
+    this.unpatch()
   }
 }