+++ packages/web/src/instrumentations/web-vitals/index.ts
@@ -25,4 +25,6 @@ export class WebVitalsInstrumentation extends BaseInstrumentation {
       });
     });
   }
+
+  shutdown(): void {}
 }